requirements_dev.txt
amqp==2.0.3
anyjson==0.3.3
backports-abc==0.4
beautifulsoup4==4.6.0
billiard==3.3.0.22
celery==3.1.23
contextlib2==0.5.5
defusedxml==0.5.0
Django==1.11.1
django-allauth==0.32.0
django-debug-toolbar==1.8
django-mptt==0.8.7
django-webpack-loader==0.5.0
djangorestframework==3.6.3
docutils==0.14
fastnumbers==0.7.4
kombu==3.0.35
lockfile==0.12.2
numpy==1.10.4
oauthlib==2.0.6
pandas==0.20.3
psycopg2==2.7.1
python-daemon==2.1.1
python-dateutil==2.5.0
python3-openid==3.1.0
pytz==2015.7
raven==6.1.0
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.8.0
rpy2==2.7.8
sendgrid-django==4.2.0
six==1.10.0
matplotlib==2.1.2
networkx==2.1
pygraphviz==1.3.1
sqlparse==0.2.4
vine==1.1.4
xmljson==0.1.9